Pull out /RecordServiceAttachmentUpdateTest of RecordServiceUploadsUpdatesDownloadsTest
This is the 2/3 patch in order to chunk down RecordServiceUploadsUpdatesDownloadsTest.
Also it makes minor changes in RecordService#updateData to increase
readability and aligns it with RecordSerice#updateData which is in the 3/3.
However this leads to a tiny behavior change - if the record does not
have an attachmentKey it resolves it.