Rewrite RecordServiceFetchModuleTest
This makes a complete rewrite of the RecordServiceFetchIntegration as module test.
However it makes minor changes of the CryptoServiceFake to fit also the
demands of RecordServiceFetchModuleTest as well as the RecordServiceModuleTestFlowHelper.
Also during the test was a minor thing found in the
RecordCompatibilityService, which had been improved alongside.
And the compatibility testset had been removed, since it not longer
needed.