Rewrite RecordServiceIntegration
This is the last patch in order to rewrite the former integration tests.
Alongside it removes the RecordServiceIntegrationBase since it is not
longer needed.
It adds downloadAttachment(s) and downloadRecord to the RecordContract
in order to make the all 3 methods accessible for the tests.
It makes final adjustments to align all Test sets regarding the module
tests of the RecordService.