Cleanup VersionList
This is a followup to the introduction of the VersionList.
VersionList does not return only a binary value any longer and utilises
now the actual status which is captured in a enum instead of individual
values.
Also it purges the access for on the VersionList from the UserService as
well as from the AndroidClient (since it was only invoked there), until
the Version model is refined.